# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API приложения для тестового задания",
        "description": "API-документация для CRUD организаций, зданий, видов деятельности и аутентификации.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "API сервер"
        }
    ],
    "paths": {
        "/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить все деятельности",
                "operationId": "b14ed4d53a740144ff049ec7cc961f2c",
                "responses": {
                    "200": {
                        "description": "List of activities"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Создать новую активность",
                "operationId": "17ca1fdfc4a2bdfc45cf306f694fa5e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Строительство"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Деятельность создана"
                    },
                    "422": {
                        "description": "Превышен максимальный уровень вложенности (3)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/activities/{id}": {
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Обновить деятельность",
                "operationId": "9cce514da02ea827e366ff51cf7bd8ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Автомобили"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Деятельность обновлена"
                    },
                    "422": {
                        "description": "Превышен максимальный уровень вложенности (3)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Удалить деятельность",
                "operationId": "37ae59b21572084eada983b39be30e04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деятельность удалена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций с фильтрами",
                "operationId": "3952cb0969bc8305f39882b2637f9b12",
                "parameters": [
                    {
                        "name": "building_id",
                        "in": "query",
                        "description": "ID здания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "activity_id",
                        "in": "query",
                        "description": "ID деятельности",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "geo_type",
                        "in": "query",
                        "description": "Тип гео-поиска (принимает значения 'box' или 'radius')",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "example": "radius"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта, если указан geo_type 'box' или 'radius'",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.7558
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота, если указан geo_type 'box' или 'radius'",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 37.6173
                        }
                    },
                    {
                        "name": "lat_2",
                        "in": "query",
                        "description": "Широта второй точки, если указан geo_type 'box'",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.7558
                        }
                    },
                    {
                        "name": "lng_2",
                        "in": "query",
                        "description": "Долгота второй точки, если указан geo_type 'box'",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 37.6173
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска в метрах (по умолчанию 1000м), если указан geo_type 'radius'",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1000
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поисковый запрос по названию организации",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "кафе"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество результатов на странице (по умолчанию 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить информацию об организации по ID",
                "operationId": "c43fd8909d9758788d18f58819842e58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации"
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Авторизация по токену Laravel Sanctum",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Activities"
        },
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}